Fix nix run with nixGL
This part of the build script was untested. This commit fixes a number
of small bug, such as pointing to the right binary path of the NixGL
package, and correctly interpolate $out in the generated script, which
was left uninterpreted (and thus would give an empty shell variable at
runtime, when $out doesn't exist anymore).